requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16.0,<3.17
djangorestframework-simplejwt>=5.5.0,<5.6
django-filter>=25.1,<26
psycopg2>=2.9.10
pillow>=11.2.1,<11.3
drf-spectacular>=0.27.2,<0.28
drf-yasg>=1.21.10,<1.22